+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## main
+
+* Fix faraday deprecation warning
+* Add `EveOnline::ESI::CharacterSendEmail`
+* Fix POST requests
+
 ## v0.42.0
 
 * Sync with new ESI